Change actions to represent what is being done.
This reduces number of actions, simplifies each action and makes
documentation easier, while at the same time increasing flexibility for
the end-user.
