+++ packages/destination-actions/src/destinations/absmartly/event.ts
@@ -1,4 +1,4 @@
-import { JSONObject, ModifiedResponse, RequestClient } from '@segment/actions-core'
+import { InputField, JSONObject, ModifiedResponse, RequestClient } from '@segment/actions-core'
 import { Settings } from './generated-types'
 import { PublishRequestUnit } from './unit'
 import { PublishRequestAttribute } from './attribute'
@@ -11,6 +11,40 @@ export interface PublishRequestEvent {
   goals?: PublishRequestGoal[]
   exposures?: JSONObject[]
   attributes?: PublishRequestAttribute[]
+}
+
+export interface DefaultPayload {
+  publishedAt: string | number
+  agent?: string
+  application?: string
+}
+
+export const defaultEventFields: Record<string, InputField> = {
+  publishedAt: {
+    label: 'Event Sent Time',
+    type: 'datetime',
+    required: true,
+    description:
+      'Exact timestamp when the event was sent (measured by the client clock). Must be an ISO 8601 date-time string, or a Unix timestamp (milliseconds) number',
+    default: {
+      '@path': '$.sentAt'
+    }
+  },
+  agent: {
+    label: 'Agent',
+    type: 'string',
+    required: false,
+    description: 'Optional agent identifier that originated the event. Used to identify which SDK generated the event.',
+    default: 'segment'
+  },
+  application: {
+    label: 'Application',
+    type: 'string',
+    required: false,
+    description:
+      'Optional application name that originated this event. Must exist if not empty. Create Applications in the Settings -> Applications section of the ABsmartly Web Console',
+    default: ''
+  }
 }
 
 export function sendEvent(
