+++ packages/destination-actions/src/destinations/absmartly/trackExposure/generated-types.ts
@@ -1,0 +1,22 @@
+// Generated file. DO NOT MODIFY IT BY HAND.
+
+export interface Payload {
+  /**
+   * The ABsmartly exposure payload without any goals. Generated by the ABsmartly SDK and should not be modified.
+   */
+  exposure: {
+    [k: string]: unknown
+  }
+  /**
+   * Exact timestamp when the event was sent (measured by the client clock). Must be an ISO 8601 date-time string, or a Unix timestamp (milliseconds) number
+   */
+  publishedAt: string | number
+  /**
+   * Optional agent identifier that originated the event. Used to identify which SDK generated the event.
+   */
+  agent?: string
+  /**
+   * Optional application name that originated this event. Must exist if not empty. Create Applications in the Settings -> Applications section of the ABsmartly Web Console
+   */
+  application?: string
+}
